Move Reports to better place
Resolves issue #16
