+++ SpiNNaker-machine/src/main/java/uk/ac/manchester/spinnaker/machine/reports/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * User-facing reports about the allocated machine.
+ *
+ * @author Donal Fellows
+ */
+package uk.ac.manchester.spinnaker.machine.reports;
